+++ packages/cubejs-playground/src/QueryBuilderV2/components/TimeListMember.tsx
@@ -1,23 +1,16 @@
-import { useRef, useState } from 'react';
-import {
-  Action,
-  Flex,
-  Space,
-  Text,
-  TimeIcon,
-  CalendarIcon,
-  TooltipProvider,
-} from '@cube-dev/ui-kit';
+import { useMemo, useRef, useState } from 'react';
+import { Flex, Space, Text, TimeIcon, TooltipProvider } from '@cube-dev/ui-kit';
 import { Cube, TCubeDimension, TimeDimensionGranularity } from '@cubejs-client/core';
 
 import { ArrowIcon } from '../icons/ArrowIcon';
 import { NonPublicIcon } from '../icons/NonPublicIcon';
 import { ItemInfoIcon } from '../icons/ItemInfoIcon';
 import { useHasOverflow } from '../hooks/has-overflow';
+import { titleize } from '../utils/index';
 
+import { GranularityListMember } from './GranularityListMember';
 import { ListMemberButton } from './ListMemberButton';
 import { FilterByMemberButton } from './FilterByMemberButton';
-import { MemberBadge } from './Badge';
 import { FilteredLabel } from './FilteredLabel';
 
 interface ListMemberProps {
@@ -32,7 +25,7 @@ interface ListMemberProps {
   onToggleDataRange?: (name: string) => void;
 }
 
-const GRANULARITIES: TimeDimensionGranularity[] = [
+const PREDEFINED_GRANULARITIES: TimeDimensionGranularity[] = [
   'second',
   'minute',
   'hour',
@@ -66,13 +59,34 @@ export function TimeListMember(props: ListMemberProps) {
   // @ts-ignore
   const description = member.description;
   const isTimestampSelected = isSelected();
-  const isGranularitySelectedList = GRANULARITIES.map((granularity) => isSelected(granularity));
-  const selectedGranularity = GRANULARITIES.find((granularity) => isSelected(granularity));
-  // const isGranularitySelected = !!isGranularitySelectedList.find((gran) => gran);
+
+  const customGranularities =
+    member.type === 'time' && member.granularities ? member.granularities.map((g) => g.name) : [];
+  const customGranularitiesTitleMap = useMemo(() => {
+    return (
+      member.type === 'time' &&
+      member.granularities?.reduce(
+        (map, granularity) => {
+          map[granularity.name] = granularity.title;
+
+          return map;
+        },
+        {} as Record<string, string>
+      )
+    );
+  }, [member.type === 'time' ? member.granularities : null]);
+  const memberGranularities = customGranularities.concat(PREDEFINED_GRANULARITIES);
+  const isGranularitySelectedMap: Record<string, boolean> = {};
+  memberGranularities.forEach((granularity) => {
+    isGranularitySelectedMap[granularity] = isSelected(granularity);
+  });
+  const selectedGranularity = memberGranularities.find((granularity) => isSelected(granularity));
 
   open = isCompact ? false : open;
 
   const hasOverflow = useHasOverflow(textRef);
+  const isAutoTitle = titleize(member.name) === title;
+
   const button = (
     <ListMemberButton
       icon={
@@ -99,22 +113,7 @@ export function TimeListMember(props: ListMemberProps) {
       <Text ref={textRef} ellipsis>
         {filterString ? <FilteredLabel text={name} filter={filterString} /> : name}
       </Text>
-      {(isCompact || !open) && selectedGranularity ? (
-        <TooltipProvider
-          delay={1000}
-          title="Click the granularity label to remove it from the query"
-        >
-          <Action
-            onPress={() => {
-              onGranularityToggle(member.name, selectedGranularity);
-            }}
-          >
-            <MemberBadge isSpecial type="timeDimension">
-              {selectedGranularity}
-            </MemberBadge>
-          </Action>
-        </TooltipProvider>
-      ) : null}
+
       <Space gap=".5x">
         <Space gap="1x">
           {description ? <ItemInfoIcon title={title} description={description} /> : undefined}
@@ -130,9 +129,35 @@ export function TimeListMember(props: ListMemberProps) {
     </ListMemberButton>
   );
 
+  const granularityItems = (items: string[], isCustom?: boolean) => {
+    return items.map((granularity: string) => {
+      if ((!open || isCompact) && !isGranularitySelectedMap[granularity]) {
+        return null;
+      }
+
+      const title = customGranularitiesTitleMap
+        ? customGranularitiesTitleMap[granularity]
+        : titleize(granularity);
+
+      return (
+        <GranularityListMember
+          key={`${name}.${granularity}`}
+          name={granularity}
+          title={title}
+          isCustom={isCustom}
+          isSelected={isGranularitySelectedMap[granularity]}
+          onToggle={() => {
+            onGranularityToggle(member.name, granularity);
+            setOpen(false);
+          }}
+        />
+      );
+    });
+  };
+
   return (
     <>
-      {hasOverflow ? (
+      {hasOverflow || !isAutoTitle ? (
         <TooltipProvider
           title={
             <>
@@ -147,7 +172,7 @@ export function TimeListMember(props: ListMemberProps) {
       ) : (
         button
       )}
-      {open || isCompact ? (
+      {open || isCompact || selectedGranularity ? (
         <Flex flow="column" gap="1bw" padding="4.5x left">
           {open && !isCompact ? (
             <ListMemberButton
@@ -162,22 +187,8 @@ export function TimeListMember(props: ListMemberProps) {
               <Text ellipsis>value</Text>
             </ListMemberButton>
           ) : null}
-          {GRANULARITIES.map((granularity, i) => {
-            return open && !isCompact ? (
-              <ListMemberButton
-                key={`${name}.${granularity}`}
-                icon={<CalendarIcon />}
-                data-member="timeDimension"
-                isSelected={isGranularitySelectedList[i]}
-                onPress={() => {
-                  onGranularityToggle(member.name, granularity);
-                  setOpen(false);
-                }}
-              >
-                <Text ellipsis>{granularity}</Text>
-              </ListMemberButton>
-            ) : null;
-          })}
+          {granularityItems(customGranularities, true)}
+          {granularityItems(PREDEFINED_GRANULARITIES)}
         </Flex>
       ) : null}
     </>