+++ rust/cubesql/cubesql/src/compile/test/test_user_change.rs
@@ -111,6 +111,36 @@ async fn test_change_user_via_in_filter_thoughtspot() {
 }
 
 #[tokio::test]
+async fn test_change_user_via_filter_powerbi() {
+    init_testing_logger();
+
+    let query_plan = convert_select_to_query_plan(
+        "SELECT COUNT(*) as cnt FROM KibanaSampleDataEcommerce WHERE NOT __user IS NULL AND __user = 'gopher'".to_string(),
+        DatabaseProtocol::PostgreSQL,
+    )
+    .await;
+
+    let cube_scan = query_plan.as_logical_plan().find_cube_scan();
+
+    assert_eq!(cube_scan.options.change_user, Some("gopher".to_string()));
+
+    assert_eq!(
+        cube_scan.request,
+        V1LoadRequestQuery {
+            measures: Some(vec!["KibanaSampleDataEcommerce.count".to_string(),]),
+            segments: Some(vec![]),
+            dimensions: Some(vec![]),
+            time_dimensions: None,
+            order: Some(vec![]),
+            limit: None,
+            offset: None,
+            filters: None,
+            ungrouped: None,
+        }
+    )
+}
+
+#[tokio::test]
 async fn test_change_user_via_filter_and() {
     let query_plan = convert_select_to_query_plan(
         "SELECT COUNT(*) as cnt FROM KibanaSampleDataEcommerce WHERE __user = 'gopher' AND customer_gender = 'male'".to_string(),
@@ -190,6 +220,47 @@ async fn test_user_with_join() {
     );
 
     assert_eq!(cube_scan.options.change_user, Some("foo".to_string()))
+}
+
+#[tokio::test]
+async fn test_change_user_via_filter_with_alias() {
+    init_testing_logger();
+
+    let query_plan = convert_select_to_query_plan(
+        r#"
+        SELECT "k"."cnt" AS "cnt"
+        FROM (
+            SELECT
+                COUNT(*) AS "cnt",
+                "__user" AS "user"
+            FROM "KibanaSampleDataEcommerce"
+            GROUP BY 2
+        ) AS "k"
+        WHERE "k"."user" = 'gopher'
+        "#
+        .to_string(),
+        DatabaseProtocol::PostgreSQL,
+    )
+    .await;
+
+    let cube_scan = query_plan.as_logical_plan().find_cube_scan();
+
+    assert_eq!(cube_scan.options.change_user, Some("gopher".to_string()));
+
+    assert_eq!(
+        cube_scan.request,
+        V1LoadRequestQuery {
+            measures: Some(vec!["KibanaSampleDataEcommerce.count".to_string(),]),
+            segments: Some(vec![]),
+            dimensions: Some(vec![]),
+            time_dimensions: None,
+            order: Some(vec![]),
+            limit: None,
+            offset: None,
+            filters: None,
+            ungrouped: None,
+        }
+    )
 }
 
 /// This should test that query with CubeScanWrapper uses proper change_user for both SQL generation and execution calls